requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
Django==6.0.3
django-filter==24.1
django-jazzmin==3.0.4
django-ninja==1.6.2
django-rest-framework==0.1.0
djangorestframework==3.17.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.31.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
